+++ sdk/containerinstance/azure-mgmt-containerinstance/generated_samples/container_groups_create_or_update.py
@@ -80,7 +80,6 @@ def main():
                     "type": "Public",
                 },
                 "osType": "Linux",
-                "provisioningTimeoutInSeconds": 600,
                 "subnetIds": [
                     {
                         "id": "[resourceId('Microsoft.Network/virtualNetworks/subnets', parameters('vnetName'), parameters('subnetName'))]"
@@ -107,6 +106,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: specification/containerinstance/resource-manager/Microsoft.ContainerInstance/preview/2023-02-01-preview/examples/ContainerGroupsCreateOrUpdate.json
+# x-ms-original-file: specification/containerinstance/resource-manager/Microsoft.ContainerInstance/stable/2023-05-01/examples/ContainerGroupsCreateOrUpdate.json
 if __name__ == "__main__":
     main()